% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count}
\alias{count}
\alias{small_count}
\title{Mock data for incomplete count data}
\format{
`count` and `small_count`
A list of y and r. The number of observations \eqn{n} is 3000 for `count` and 100 for `small_count`.
\describe{
  \item{y}{A counr variable. It contains missing entries.}
  \item{r}{A binary indicator variable. \eqn{R_i = 1} when \eqn{Y_i} is observed.}
}

An object of class \code{data.frame} with 100 rows and 2 columns.
}
\usage{
count

small_count
}
\description{
Data generated for inference with incomplete count data
}
\keyword{datasets}
