% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross}
\alias{cross}
\alias{small_cross}
\title{Mock data for cross-sectional studies}
\format{
`cross` and `small_cross`
A vector with 4 entries of matrix \eqn{y}, total count is 38000 for `cross` and 38 for `small_cross`:
\describe{
  \item{element 1}{\eqn{y_{1,1}}}
  \item{element 2}{\eqn{y_{1,2}}}
  \item{element 3}{\eqn{y_{2,1}}}
  \item{element 4}{\eqn{y_{2,2}}}
}

An object of class \code{integer} of length 4.
}
\usage{
cross

small_cross
}
\description{
Data generated for cross-sectional studies
}
\keyword{datasets}
