% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_identification}
\alias{diagnose_identification}
\title{Diagnosing the identification of the model}
\usage{
diagnose_identification(jags_code, data, variable.names, ll_fun, data_fun)
}
\arguments{
\item{jags_code}{string. JAGS model code}

\item{data}{list. Data list of either vector or number}

\item{variable.names}{vector of string. Names of the variables}

\item{ll_fun}{function. Log-likelihood function that takes
two arguments - data and theta (parameters)}

\item{data_fun}{function. Function that generates the dataset
using theta (parameters).
argument of this function is
a list of arguments required for creating the dataset.}
}
\value{
list. A list of the condition number, rank, and asymptotic variances.
}
\description{
Estimate parameters using small posterior samples from JAGS,
calculate the approximate Fisher information matrix
and return the condition number and rank of the approximate Fisher information matrix
with asymptotic variances of estimated parameters.
}
\examples{
ll_cross <- function(data, theta){
se = theta[1]
sp = theta[2]
pi = theta[3:5]
pi = c(pi, 1-sum(pi))
eta =  c(se*pi[1]+(1-sp)*pi[3],
  se*pi[2]+(1-sp)*pi[4],
  sp*pi[3]+(1-se)*pi[1],
  (1-se)*pi[2]+sp*pi[4])
  log(dmultinom(data, prob = eta))
  }

  data_cross <- function(args) {
    theta = args$theta
    n = args$n
    se = theta[1]
    sp = theta[2]
    pi = theta[3:5]
    pi = c(pi, 1-sum(pi))
    eta =  c(se*pi[1]+(1-sp)*pi[3],
      se*pi[2]+(1-sp)*pi[4],
      sp*pi[3]+(1-se)*pi[1],
      (1-se)*pi[2]+sp*pi[4])
      as.vector(rmultinom(1, size = n, prob = eta))
     }
jags_code <-
"model{
y[1:4] ~ dmulti(eta[1:4], n)
eta[1:4] <- c(se*pi[1]+(1-sp)*pi[3],
se*pi[2]+(1-sp)*pi[4],
sp*pi[3]+(1-se)*pi[1],
(1-se)*pi[2]+sp*pi[4])
pi[1:4] ~ ddirch(c(1,1,1,1))
se ~ dbeta(25,3)
sp ~ dbeta(30,1.5)
}"
data(small_cross)
data = list(y = small_cross, n = sum(small_cross))
diagnose_identification(jags_code, data, c("se","sp","pi[1]","pi[2]","pi[3]"),ll_cross,data_cross)

}
