% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_julia_session_multi}
\alias{init_julia_session_multi}
\title{Initialize Julia session}
\usage{
init_julia_session_multi(threads)
}
\arguments{
\item{threads}{The number of threads used in Julia session}
}
\description{
It needs to be required to run before any functions in this package
The number of threads in Julia can't be changed once session starts.
To restart Julia with new number of threads, restart of R session and run this function.
}
