% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postInference.R
\name{pigeons_plots}
\alias{pigeons_plots}
\title{Trace and density plot from pigeons results}
\usage{
pigeons_plots(pt)
}
\arguments{
\item{pt}{pigeons results from sampling function}
}
\description{
This function returns the plots from pigeons results.
}
