% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postInference.R
\name{pigeons_samples}
\alias{pigeons_samples}
\title{Extract samples from pigeons results}
\usage{
pigeons_samples(pt)
}
\arguments{
\item{pt}{pigeons results from sampling function}
}
\value{
a matrix of samples
}
\description{
This function returns the matrix of samples from pigeons results.
}
