% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sat}
\alias{sat}
\alias{small_sat}
\title{Mock data for incomplete binary data}
\format{
`sat` and `small_sat`
A list of x, y, and r. The number of observations \eqn{n} is 3000 for `sat` and 100 for `small_sat`.
The number of predictors \eqn{p} in \eqn{X} is 4 for `sat` and 2 for `small_sat.`
\describe{
  \item{x}{A design matrix of X, binary. \eqn{2^p} columns and \eqn{n} rows.}
  \item{y}{A binary response variable. It contains missing entries.}
  \item{r}{A binary indicator variable. \eqn{R_i = 1} when \eqn{Y_i} is observed.}
}

An object of class \code{list} of length 3.
}
\usage{
sat

small_sat
}
\description{
Data generated for inference with incomplete binary data
}
\keyword{datasets}
