% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{setup}
\alias{setup}
\title{Initial setup function.}
\usage{
setup(installJulia = FALSE, JULIA_HOME = NULL)
}
\arguments{
\item{installJulia}{boolean if TRUE, it will install Julia}

\item{JULIA_HOME}{(Optional) string indicating the location of specific Julia binary}
}
\description{
Set up Julia environment, and install Julia if necessary.
Install required Julia packages.
}
